requirements.txt
# BASE
rq>=2.6.0
redis>=6.4.0  # Updated from 5.2.1 to 6.4.0 (merged with main)
pandas>=2.2.3
transformers>=4.56.1
speakerpy>=0.1.2
pyside6>=6.9.0
outcome>=1.3.0.post0
trio>=0.29.0
loguru>=0.7.3
setuptools>=80.0.0
pynput>=1.8.1
pyserial>=3.5

# Speech processing
vosk>=0.3.45
speechbrain>=0.5.16
silero>=0.0.1
soundfile>=0.12.1
pyaudio>=0.2.14
numpy>=1.24.0
torch>=2.0.0
openai-whisper>=20231117

# Face detection and recognition
opencv-python>=4.8.0
face-recognition>=1.3.0
dlib>=19.24.0

# DEV dependencies (install separately: pip install -r requirements-dev.txt)
# mypy>=1.17.1
# ruff>=0.12.12
# black>=25.1.0
# pytest>=8.4.2